Rename ipam.NetworkAllocator service -> network.Allocator service
I feel like there's too many ipam packages now and it makes writing unambiguous
code a bit painful. I feel like network is a better package name here and also
rhimes with allocator better.
